pyproject.toml
[tool.poetry]
name = "cloud-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.115.0"}
psycopg2-binary = "^2.9.9"
alembic = "^1.13.3"
sqlalchemy = "^2.0.35"
pydantic-settings = "^2.5.2"
typer = "^0.12.5"
pytest = "^8.3.3"
coverage = "^7.6.2"
pytest-cov = "^5.0.0"
black = "^24.10.0"
isort = "^5.13.2"
pylint = "^3.3.1"
celery = {extras = ["redis"], version = "^5.4.0"}
strenum = "^0.4.15"
google-cloud-pubsub = "^2.26.0"
pyjwt = "^2.9.0"
locust = "^2.31.8"
sqlalchemy-json = "^0.7.0"
pre-commit = "^4.0.1"
faker = "^30.3.0"
boto3 = "^1.35.39"
moviepy = "^1.0.3"
pillow = "9.5.0"
cryptography = "^43.0.1"
pycryptodome = "^3.21.0"
google-cloud-storage = "^2.18.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.coverage.run]
branch = true
data_file = ".coverage"
omit = ["*/main.py", "tests/*", "*/pubsub_worker.py", "*/celery_worker.py", "*/settings"]

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.report]
fail_under = 50

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
# log_cli = true
